236 fix 5차 테스트 에러

+++ src/main/java/com/floney/floney/analyze/dto/response/AnalyzeResponseByBudget.java
@@ -4,19 +4,19 @@ import lombok.Getter;
 
 @Getter
 public class AnalyzeResponseByBudget {
-    private final long leftMoney;
-    private final long initBudget;
+    private final float leftMoney;
+    private final float initBudget;
 
-    private AnalyzeResponseByBudget(long totalOutcomeMoney, long initBudget) {
+    private AnalyzeResponseByBudget(float totalOutcomeMoney, float initBudget) {
         this.initBudget = initBudget;
         this.leftMoney = calculateLeftMoney(initBudget, totalOutcomeMoney);
     }
 
-    public static AnalyzeResponseByBudget of(Long totalOutcome, Long initBudget) {
+    public static AnalyzeResponseByBudget of(float totalOutcome, float initBudget) {
         return new AnalyzeResponseByBudget(totalOutcome, initBudget);
     }
 
-    private long calculateLeftMoney(long initBudget, long totalOutcomeMoney) {
+    private float calculateLeftMoney(float initBudget, float totalOutcomeMoney) {
         return initBudget - totalOutcomeMoney;
     }
 }
